Fixed the missing code

+++ Frontend/src/RockRoute/Models/Climbs.cs
@@ -21,7 +21,7 @@ namespace RockRoute.Models //makes accessible from other areas of the project
         public required climbTypes Type {get; set;}
 
         public required string YDS {get; set;}
-        public required (float Lat, float Long) ParentLocation {get; set;}
+        public required (double Lat, double Long) ParentLocation {get; set;}
         public required string LocationDescription {get; set;}
         public required string Protection_Notes {get; set;}
         public required List<CRating> UserRatings {get; set;} //List of Tuples with (String, int)
